Crate problem122

Problem 122 (Efficient exponentiation)

Solution detail

This is, sadly, a problem where a fully correct solution will take much longer to run than we'd ideally be happy with. But we can take a shortcut!

Firstly, it is clear by considering the point of view of the exponents that we are looking for optimal Addition Chains. Furthermore, that link explains that, for all n < 12509, there is an optimal addition chain which is also a star chain - that is, each addition uses the most recent result.

We can therefore vastly cut down the search space by only considering star chains. That makes a traditional depth-first search, where each node consists of the current power and a list of all previous powers, a suitable method for finding all optimal star chains.

Keep track, as the search progresses, of the shortest chain found for each power, and sum them after the search is complete.

Structs

StarChainTree
The state of the search for star chains

StarChainTreeStep
A step that can be taken in the search tree

Functions

answer
Solve the problem, returning the answer as a String

main

solve
Find the total length of all optimal star chains with targets up to the given limit.